Re-organize the code into packages.
