+++ app/src/androidTest/java/ch/tarsier/tarsier/test/ProfileActivityTest.java
@@ -2,7 +2,7 @@ package ch.tarsier.tarsier.test;
 
 import android.test.ActivityInstrumentationTestCase2;
 
-import ch.tarsier.tarsier.ProfileActivity;
+import ch.tarsier.tarsier.ui.activity.ProfileActivity;
 import ch.tarsier.tarsier.R;
 
 import static ch.tarsier.tarsier.test.matchers.HasErrorMatcher.hasError;
