+++ app/src/main/AndroidManifest.xml
@@ -33,7 +33,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name=".HomeActivity"
+            android:name=".ui.activity.HomeActivity"
             android:label="@string/app_name" >
 
             <!-- android:theme="@android:style/Theme.Holo.Light.NoActionBar" -->
@@ -44,38 +44,38 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".AddProfilePictureActivity"
+            android:name=".ui.activity.AddProfilePictureActivity"
             android:label="@string/title_activity_add_profile_picture"
             android:theme="@style/AppTheme" >
         </activity>
         <activity
-            android:name=".ProfileActivity"
+            android:name=".ui.activity.ProfileActivity"
             android:label="@string/title_activity_profile" >
         </activity>
         <activity
-            android:name=".WiFiDirectDebugActivity"
+            android:name=".network.old.WiFiDirectDebugActivity"
             android:label="@string/title_activity_wifidirectdebug" >
         </activity>
         <activity
-            android:name=".NewDiscussionActivity"
+            android:name=".ui.activity.NewDiscussionActivity"
             android:label="@string/title_activity_new_discussion" >
         </activity>
         <activity
-            android:name=".ChatRoomParticipantsActivity"
+            android:name=".ui.activity.ChatRoomParticipantsActivity"
             android:label="@string/title_activity_chat_room_participants" >
         </activity>
         <activity
-            android:name=".ui.PreferencesActivity"
+            android:name=".ui.activity.PreferencesActivity"
             android:label="@string/title_activity_preferences" >
         </activity>
 
-        <activity android:name=".DiscussionsActivity"
+        <activity android:name=".ui.activity.DiscussionsActivity"
             android:label="Discussions"
             android:theme="@style/CustomActionBarTheme" >
         </activity>
 
         <activity
-            android:name=".ConversationActivity"
+            android:name=".ui.activity.ConversationActivity"
             android:label="@string/title_activity_conversation" >
         </activity>
     </application>
