+++ app/src/main/java/ch/tarsier/tarsier/network/old/ChatRoom.java
@@ -1,4 +1,4 @@
-package ch.tarsier.tarsier;
+package ch.tarsier.tarsier.network.old;
 
 import android.app.Fragment;
 import android.content.Context;
@@ -13,6 +13,8 @@ import android.widget.TextView;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import ch.tarsier.tarsier.R;
 
 /**
  * Created by amirreza on 10/27/14.
