+++ app/src/main/java/ch/tarsier/tarsier/network/old/ConnectionListener.java
