+++ app/src/main/java/ch/tarsier/tarsier/network/old/WiFiDirectBroadcastReceiver.java
@@ -1,4 +1,4 @@
-package ch.tarsier.tarsier;
+package ch.tarsier.tarsier.network.old;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
