+++ app/src/main/java/ch/tarsier/tarsier/storage/DiscussionSummary.java
@@ -1,4 +1,4 @@
-package ch.tarsier.tarsier;
+package ch.tarsier.tarsier.storage;
 
 /**
  * Created by gluthier
