+++ app/src/main/java/ch/tarsier/tarsier/storage/MessageViewModel.java
@@ -1,4 +1,4 @@
-package ch.tarsier.tarsier;
+package ch.tarsier.tarsier.storage;
 
 
 import android.graphics.Bitmap;
