+++ app/src/main/java/ch/tarsier/tarsier/ui/activity/DiscussionsActivity.java
@@ -1,7 +1,6 @@
-package ch.tarsier.tarsier;
+package ch.tarsier.tarsier.ui.activity;
 
 import android.app.Activity;
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -9,6 +8,10 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.Toast;
+
+import ch.tarsier.tarsier.storage.DiscussionSummary;
+import ch.tarsier.tarsier.ui.adapter.DiscussionsAdapter;
+import ch.tarsier.tarsier.R;
 
 /**
  * Created by gluthier
