+++ app/src/main/java/ch/tarsier/tarsier/ui/activity/HomeActivity.java
@@ -1,4 +1,4 @@
-package ch.tarsier.tarsier;
+package ch.tarsier.tarsier.ui.activity;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -16,7 +16,8 @@ import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.Toast;
 
-import ch.tarsier.tarsier.storage.ChatsContract;
+import ch.tarsier.tarsier.R;
+import ch.tarsier.tarsier.network.old.WiFiDirectDebugActivity;
 import ch.tarsier.tarsier.validation.StatusMessageValidator;
 import ch.tarsier.tarsier.validation.UsernameValidator;
