+++ app/src/main/java/ch/tarsier/tarsier/ui/adapter/DiscussionsAdapter.java
@@ -1,4 +1,4 @@
-package ch.tarsier.tarsier;
+package ch.tarsier.tarsier.ui.adapter;
 
 import android.app.Activity;
 import android.content.Context;
@@ -8,6 +8,9 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import ch.tarsier.tarsier.storage.DiscussionSummary;
+import ch.tarsier.tarsier.R;
 
 /**
  * Created by gluthier
