+++ app/src/main/java/ch/tarsier/tarsier/ui/view/EndlessListener.java
@@ -1,4 +1,4 @@
-package ch.tarsier.tarsier;
+package ch.tarsier.tarsier.ui.view;
 
 /**
  * @author xawill
