+++ app/src/main/res/layout/activity_add_profile_picture.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="ch.tarsier.tarsier.AddProfilePictureActivity">
+    tools:context="ch.tarsier.tarsier.ui.activity.AddProfilePictureActivity">
 
     <RelativeLayout
         android:layout_width="fill_parent"
