+++ app/src/main/res/layout/activity_conversation.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ch.tarsier.tarsier.ConversationActivity"
+    tools:context="ch.tarsier.tarsier.ui.activity.ConversationActivity"
     android:orientation="vertical" >
 
     <EndlessListView
