+++ app/src/main/res/layout/activity_new_discussion.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="ch.tarsier.tarsier.NewDiscussionActivity" >
+    tools:context="ch.tarsier.tarsier.ui.activity.NewDiscussionActivity" >
 
     <EditText
         android:layout_width="fill_parent"
