+++ app/src/main/res/layout/activity_profile.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="ch.tarsier.tarsier.ProfileActivity"
+    tools:context="ch.tarsier.tarsier.ui.activity.ProfileActivity"
     android:id="@+id/rel_layout">
 
     <ImageView
