+++ app/src/main/res/menu/add_profile_picture.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="ch.tarsier.tarsier.AddProfilePictureActivity" >
+    tools:context="ch.tarsier.tarsier.ui.activity.AddProfilePictureActivity" >
 
     <item android:id="@+id/action_settings"
         android:title="@string/action_settings"
