+++ app/src/main/res/menu/chat_room_participants.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="ch.tarsier.tarsier.ChatRoomParticipantsActivity" >
+    tools:context="ch.tarsier.tarsier.ui.activity.ChatRoomParticipantsActivity" >
     <item android:id="@+id/action_settings"
         android:title="@string/action_settings"
         android:orderInCategory="100"
