+++ app/src/main/res/menu/new_discussion.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="ch.tarsier.tarsier.NewDiscussionActivity" >
+    tools:context="ch.tarsier.tarsier.ui.activity.NewDiscussionActivity" >
 
     <item android:id="@+id/create_room"
         android:icon="@drawable/check_profile"
