+++ app/src/main/res/menu/preferences.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="ch.tarsier.tarsier.ui.PreferencesActivity" >
+    tools:context="ch.tarsier.tarsier.ui.activity.PreferencesActivity" >
