+++ app/src/main/res/menu/private_discussion.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="ch.tarsier.tarsier.ConversationActivity" >
+    tools:context="ch.tarsier.tarsier.ui.activity.ConversationActivity" >
 
     <item android:id="@+id/action_settings"
         android:title="@string/action_settings"
