+++ app/src/main/res/menu/profile.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="ch.tarsier.tarsier.ProfileActivity" >
+    tools:context="ch.tarsier.tarsier.ui.activity.ProfileActivity" >
     <item android:id="@+id/action_save_profile"
           android:icon="@drawable/check_profile"
           android:title="@string/save_profile"
